requirements.txt
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.78.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
kombu==5.2.4
loguru==0.6.0
packaging==21.3
pika==1.2.1
prompt-toolkit==3.0.30
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pysnowflake==0.1.3
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tornado==6.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
